Develop

+++ SyudentAccounting.BusinessLogic/Services/Implementations/ApplicationsInTheProjectService.cs
@@ -15,27 +15,61 @@ namespace StudentAccounting.BusinessLogic.Services.Implementations
         }
         public void Create(ApplicationsInTheProject applicationsInTheProject)
         {
+            if (applicationsInTheProject == null)
+            {
+                throw new ArgumentNullException(nameof(applicationsInTheProject));
+            }
+            if (_context.ApplicationsInTheProjects.Any(x => x.Id == applicationsInTheProject.Id))
+            {
+                throw new Exception();
+            }
             _context.ApplicationsInTheProjects.Add(applicationsInTheProject);
             _context.SaveChanges();
         }
         public IEnumerable<ApplicationsInTheProject> Get()
         {
-            return _context.ApplicationsInTheProjects.Include(x=>x.Vacancy).Include(x=> x.Participants).AsNoTracking().ToList();
+            return _context.ApplicationsInTheProjects.Include(x => x.Vacancy).Include(x => x.Participants).AsNoTracking().ToList();
         }
         public ApplicationsInTheProject Get(int id)
         {
-            return _context.ApplicationsInTheProjects.Include(x => x.Vacancy).Include(x => x.Participants).AsNoTracking().FirstOrDefault(x => x.Id == id);
+            if (id <= 0)
+            {
+                throw new Exception();
+            }
+            var application = _context.ApplicationsInTheProjects.Include(x => x.Vacancy).Include(x => x.Participants).AsNoTracking().FirstOrDefault(x => x.Id == id);
+            if (application == null)
+            {
+                throw new Exception();
+            }
+            return application;
         }
         public void Edit(ApplicationsInTheProject applicationsInTheProject)
         {
+            if (applicationsInTheProject == null)
+            {
+                throw new ArgumentNullException(nameof(applicationsInTheProject));
+            }
+            if (!_context.ApplicationsInTheProjects.Any(x => x.Id == applicationsInTheProject.Id))
+            {
+                throw new Exception();
+            }
             _context.ApplicationsInTheProjects.Update(applicationsInTheProject);
             _context.SaveChanges();
         }
         public void Delete(int id)
         {
+            if (id <= 0)
+            {
+                throw new ArgumentOutOfRangeException(nameof(id));
+            }
             var application = _context.ApplicationsInTheProjects.FirstOrDefault(x => x.Id == id);
+            if (application == null)
+            {
+                throw new Exception();
+            }
             _context.ApplicationsInTheProjects.Remove(application);
             _context.SaveChanges();
         }
     }
 }
+
